refactor(api-gateway): modularize webhook and tasks
