+++ apps/api-gateway/src/app.ts
@@ -3,13 +3,13 @@ import 'reflect-metadata';
 import express from 'express';
 import dotenv from 'dotenv';
 import { container } from 'tsyringe';
-import cron from 'node-cron';
-import fetch from 'node-fetch';
 import logger from './services/logger-service.js';
+import { setupWhatsAppWebhook } from './whatsapp-webhook.js';
+import { setupScheduledTasks } from './scheduled-tasks.js';
+import { setupGracefulShutdown } from './graceful-shutdown.js';
 
 // Import services and configurations
-import { Database, users } from '@running-coach/database';
-import { eq } from 'drizzle-orm';
+import { Database } from '@running-coach/database';
 import { ChatBuffer, VectorMemory } from '@running-coach/vector-memory';
 
 import { LanguageDetector, TemplateEngine, I18nService } from '@running-coach/shared';
@@ -22,42 +22,6 @@ import { HeadCoach, LLMClient } from '@running-coach/llm-orchestrator';
 
 // Load environment variables
 dotenv.config();
-
-// Function to send WhatsApp messages via Meta API
-async function sendWhatsAppMessage(to: string, message: string, config: any) {
-  try {
-    const url = `https://graph.facebook.com/v17.0/${config.NUMBER_ID}/messages`;
-    
-    const response = await fetch(url, {
-      method: 'POST',
-      headers: {
-        'Authorization': `Bearer ${config.JWT_TOKEN}`,
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify({
-        messaging_product: 'whatsapp',
-        to: to,
-        type: 'text',
-        text: {
-          body: message
-        }
-      })
-    });
-
-    if (!response.ok) {
-      const errorText = await response.text();
-      logger.error(`❌ Failed to send WhatsApp message: ${response.status} - ${errorText}`);
-      return false;
-    }
-
-    const result = await response.json();
-    logger.info(`✅ WhatsApp message sent successfully:`, result);
-    return true;
-  } catch (error) {
-    logger.error('❌ Error sending WhatsApp message:', error);
-    return false;
-  }
-}
 
 // Configuration interfaces
 interface Config {
@@ -251,35 +215,6 @@ async function initializeServices(config: Config) {
 // Initialize WhatsApp Bot
 
 
-// Setup scheduled tasks
-function setupScheduledTasks(services: any) {
-  logger.info('⏰ Setting up scheduled tasks...');
-
-  // Generate weekly progress summaries every Sunday at 9 AM
-  cron.schedule('0 9 * * 0', async () => {
-    logger.info('📊 Running weekly progress summary generation...');
-    try {
-      // This would typically get all active users and generate summaries
-      // For now, we'll implement a placeholder
-      logger.info('Weekly progress summaries generated');
-    } catch (error) {
-      logger.error('Error generating weekly summaries:', error);
-    }
-  });
-
-  // Daily VDOT recalculation at midnight
-  cron.schedule('0 0 * * *', async () => {
-    logger.info('🧮 Running daily VDOT recalculation...');
-    try {
-      // Recalculate VDOT for users with new runs
-      logger.info('Daily VDOT recalculation completed');
-    } catch (error) {
-      logger.error('Error in daily VDOT recalculation:', error);
-    }
-  });
-
-  logger.info('✅ Scheduled tasks configured');
-}
 
 // Setup health endpoints
 function setupHealthEndpoints(app: express.Application, services: any) {
@@ -334,25 +269,6 @@ running_coach_users_active 0
   });
 }
 
-// Graceful shutdown
-function setupGracefulShutdown(services: any) {
-  const shutdown = async (signal: string) => {
-    console.log(`\n🛑 Received ${signal}, shutting down gracefully...`);
-    
-    try {
-      await services.database.close();
-      await services.chatBuffer.close();
-      console.log('✅ All services closed successfully');
-      process.exit(0);
-    } catch (error) {
-      console.error('❌ Error during shutdown:', error);
-      process.exit(1);
-    }
-  };
-
-  process.on('SIGINT', () => shutdown('SIGINT'));
-  process.on('SIGTERM', () => shutdown('SIGTERM'));
-}
 
 // Main application
 async function main() {
@@ -368,114 +284,8 @@ async function main() {
     const app = express();
     // Parse JSON bodies for incoming requests
     app.use(express.json());
-    
-    // Configure WhatsApp webhook endpoint
-    app.get('/webhook', (req, res) => {
-      const mode = req.query['hub.mode'] as string;
-      const token = req.query['hub.verify_token'] as string;
-      const challenge = req.query['hub.challenge'] as string;
-      
-      // Check if a token and mode is in the query string of the request
-      if (mode && token) {
-        // Check the mode and token sent is correct
-        if (mode === 'subscribe' && token === config.VERIFY_TOKEN) {
-          // Respond with the challenge token from the request
-          logger.info('✅ WEBHOOK_VERIFIED');
-          res.status(200).send(challenge);
-        } else {
-          // Respond with '403 Forbidden' if verify tokens do not match
-          logger.warn('❌ WEBHOOK_VERIFICATION_FAILED: Token mismatch');
-          res.sendStatus(403);
-        }
-      } else {
-        logger.warn('❌ WEBHOOK_VERIFICATION_FAILED: Missing parameters');
-        res.sendStatus(400);
-      }
-    });
-    
-    // Configure webhook POST endpoint for receiving messages
-    app.post('/webhook', async (req, res) => {
-      // Always respond with 200 OK to WhatsApp as required
-      res.status(200).send('OK');
-      
-      try {
-        // Process the incoming webhook data
-        const data = req.body;
-        logger.info('💬 Received WhatsApp webhook:', JSON.stringify(data, null, 2));
-        
-        // Process webhook data for WhatsApp Business API
-        if (data && data.object === 'whatsapp_business_account') {
-          logger.info('💬 Processing WhatsApp webhook data...');
-          
-          // Extract messages and process them
-          if (data.entry && Array.isArray(data.entry)) {
-            for (const entry of data.entry) {
-              if (entry.changes && Array.isArray(entry.changes)) {
-                for (const change of entry.changes) {
-                  if (change.value && change.value.messages && Array.isArray(change.value.messages)) {
-                    logger.info(`📱 Found ${change.value.messages.length} messages to process`);
-                    
-                    // Process each message with AI Agent
-                    for (const message of change.value.messages) {
-                      try {
-                        if (message.type === 'text' && message.text && message.text.body) {
-                          const phone = message.from;
-                          const messageText = message.text.body;
 
-                          logger.info(`📩 Processing message from ${phone}: ${messageText}`);
-
-                          // Get or create user
-                          let [user] = await services.database.query
-                            .select()
-                            .from(users)
-                            .where(eq(users.phoneNumber, phone))
-                            .limit(1);
-                          if (!user) {
-                            [user] = await services.database.query
-                              .insert(users)
-                              .values({ phoneNumber: phone, preferredLanguage: 'es' })
-                              .returning();
-                          }
-
-                          const allowance = await services.freemiumService.checkMessageAllowance(user);
-                          if (!allowance.allowed) {
-                            const upsell = user.preferredLanguage === 'en'
-                              ? `You reached the free message limit. Upgrade here: ${allowance.link}`
-                              : `Has alcanzado tu límite gratuito. Mejora aquí: ${allowance.link}`;
-                            await sendWhatsAppMessage(phone, upsell, config);
-                            continue;
-                          }
-
-                          // Process message with Head Coach
-                          const aiResponse = await services.headCoach.handleMessage({
-                            userId: user.id,
-                            conversationHistory: [], // Simplified for now
-                            userMessage: messageText,
-                            channel: 'whatsapp',
-                          });
-
-                          if (aiResponse && aiResponse.length > 0) {
-                            // Send response back to WhatsApp
-                            await sendWhatsAppMessage(phone, aiResponse, config);
-                            logger.info(`✅ Sent AI response to ${phone}: ${aiResponse}`);
-                          }
-                        }
-                      } catch (messageError) {
-                        logger.error('❌ Error processing individual message:', messageError);
-                      }
-                    }
-                  }
-                }
-              }
-            }
-          }
-        }
-      } catch (error) {
-        logger.error('❌ Error processing webhook:', error);
-      }
-    });
-
-    
+    setupWhatsAppWebhook(app, services, config);
     setupHealthEndpoints(app, services);
     setupScheduledTasks(services);
     setupGracefulShutdown(services);
